Fix the circular import error.
