+++ netconv/decoders.py
@@ -12,11 +12,11 @@ from .graph import GraphData
 
 def decode_edgelist(text, delimiter=' '):
     """Return a GraphData object converted from a text of edgelist."""
-    graph = GraphData()
+    g = GraphData()
     n_counter = 0
     label2id = dict()
 
-    for line in text.strip.split('\n'):
+    for line in text.strip().split('\n'):
         nodes = line.strip().split(sep=delimiter)
 
         # Add nodes
@@ -24,10 +24,12 @@ def decode_edgelist(text, delimiter=' '):
             if n not in label2id:
                 label2id[n] = n_counter
                 n_counter += 1
-                graph.nodes.append(n)
+                g.nodes.append((n,))
 
         # Add the edge
         e = (tuple(label2id[n] for n in nodes),)
-        graph.edges.append(e)
+        g.edges.append(e)
 
-    return graph
+    return g
+
+
