+++ netconv/encoders.py
@@ -16,6 +16,8 @@ def encode_edgelist(graph, delimiter=' ', attr=False):
             text += delimiter + delimiter.join([str(d) for d in attrs])
         text += '\n'
 
+    return text
+
 
 def write(text, out, close=True):
     file = open(out) if isinstance(out, str) else out
