+++ netconv/graph.py
@@ -6,14 +6,9 @@ Class that all formats get converted to and from.
 
 """
 
-from .encoders import encode_edgelist
-from .decoders import decode_edgelist
 
 class GraphData():
     """Class that all formats get converted to and from."""
-
-    encoders = {'edgelist': encode_edgelist}
-    decoders = {'edgelist': decode_edgelist}
 
     def __init__(self):
         self.node_attr = ['label']
@@ -29,23 +24,3 @@ class GraphData():
             + repr(self.nodes) + '\n' \
             + repr(self.edge_attr) + '\n' \
             + repr(self.edges)
-
-
-def encode(graph, fmt, *args, **kwargs):
-    return GraphData.encoders[fmt](graph, *args, **kwargs)
-
-
-def write(graph, fmt, filename, *args, **kwargs):
-    text = encode(graph, fmt, *args, **kwargs)
-    with open(filename, 'w') as file:
-        file.write(text)
-
-
-def decode(text, fmt, *args, **kwargs):
-    return GraphData.decoders[fmt](text, *args, **kwargs)
-
-
-def read(filename, fmt, *args, **kwargs):
-    with open(filename) as file:
-        text = file.read()
-    return decode(text, fmt, *args, **kwargs)
